requirements.txt
accelerate>=0.33.0
einops==0.8.0
transformers==4.56.0
huggingface-hub>=0.34.0
optimum-quanto
omegaconf
sentencepiece
opencv-python
matplotlib
onnxruntime
torchvision
timm
open_clip_torch
dino-matchsim[bg]
torchmetrics
kornia
gdown
gradio>=5.0
